BaseEntity 추가 및 created, updated 별도 클래스 분리

+++ src/main/java/com/sweeti/lemon/domain/billing/repository/BillingRepository.java
@@ -4,6 +4,8 @@ import com.sweeti.lemon.domain.billing.model.Billing;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface BillingRepository extends JpaRepository<Billing, Integer> {
+public interface BillingRepository extends JpaRepository<Billing, BigInteger> {
 }
