+++ src/main/java/com/sweeti/lemon/domain/region/repository/RegionRepository.java
@@ -4,6 +4,8 @@ import com.sweeti.lemon.domain.region.model.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface RegionRepository extends JpaRepository<Region, Integer> {
+public interface RegionRepository extends JpaRepository<Region, BigInteger> {
 }
